+++ lib/hexapdf/configuration.rb
@@ -505,6 +505,7 @@ module HexaPDF
                         },
                         XXAcroFormField: {
                           Tx: 'HexaPDF::Type::AcroForm::TextField',
+                          Btn: 'HexaPDF::Type::AcroForm::ButtonField',
                         },
                       })